refactor : Ticket 리팩터링
* Ticket의 생성자를 두개로 줄임.
* 기존의 split하던 기능은 util의 generator로 따로 분리
